bugfix: target lock acquisition out of order
To fix this, cleaned up the outputs list in build_if_required which now always
contains the target.

It is no longer passed through build to run_recipe for rename-on-error, which
now instead gets the "side" outputs from the Production.

Accessing the Production was made possible by an independently motivated
refactoring that provides for a nicer object-oriented abstraction: _run_recipe
(as well as _build and build_if_required) is now a method of a Producer "actor"
class (formerly known as BuildIfRequiredThread).