Rename-on-error now done as a global cleanup task.
This is more robust; workers may hang when their subprocess (which also
receives the SIGINT) does not exit, thus when they were doing the cleanup, this
might happen at an unspecified time in the future, or never.

Also, removed interrupt handling code from workers - KeyboardInterrupt is only
raised in main thread, but the subprocess also receives the SIGINT and normally
should exit with non-0, so then to the worker threads looks like a regular
recipe failure.