Bugfix: set up logging before logger is accessed
Status message logging is now setup when the produce module is loaded rather
than after the produce method is first called.

Because setup removes all handlers from the produce loggers (to get rid of the
default handler - is there a better way???), if it ran *within* an assertLogs
block, the test would fail. The bug was elusive because running all tests still
succeeded: setup already happened before the assertLogs test. Only when running
such a test in isolation did the bug surface.