Documented sideways dependencies.
This seems complicated and it is tempting to think we would be better off not
declaring outputs withing the rules that create them but in "proxy" rules,
which we need to declare and specially support (except to the soft cycle ban).

But ATM we do need to declare the outputs, we use them for locking and for
pretend mode.